+++ auth/presentation/src/main/kotlin/me/proton/core/auth/presentation/ui/ChooseAddressActivity.kt
@@ -34,6 +34,7 @@ import me.proton.core.auth.presentation.entity.CreateAddressInput
 import me.proton.core.auth.presentation.entity.CreateAddressResult
 import me.proton.core.auth.presentation.viewmodel.ChooseAddressViewModel
 import me.proton.core.domain.entity.UserId
+import me.proton.core.presentation.utils.getUserMessage
 import me.proton.core.presentation.utils.hideKeyboard
 import me.proton.core.presentation.utils.onClick
 import me.proton.core.presentation.utils.onFailure
@@ -86,7 +87,7 @@ class ChooseAddressActivity : AuthActivity<ActivityChooseAddressBinding>(Activit
                 is ChooseAddressViewModel.State.Processing -> showLoading(true)
                 is ChooseAddressViewModel.State.Success -> onUsernameAvailable(it.username, it.domain)
                 is ChooseAddressViewModel.State.Data -> onData(it.username, it.domains)
-                is ChooseAddressViewModel.State.Error.Message -> showError(it.message)
+                is ChooseAddressViewModel.State.Error.Message -> showError(it.error.getUserMessage(resources))
                 is ChooseAddressViewModel.State.Error.DomainsNotAvailable ->
                     showError(getString(R.string.auth_create_address_error_no_available_domain))
                 is ChooseAddressViewModel.State.Error.UsernameNotAvailable ->