+++ auth/presentation/src/main/kotlin/me/proton/core/auth/presentation/ui/SecondFactorActivity.kt
@@ -36,6 +36,7 @@ import me.proton.core.auth.presentation.entity.SecondFactorResult
 import me.proton.core.auth.presentation.viewmodel.SecondFactorViewModel
 import me.proton.core.domain.entity.UserId
 import me.proton.core.presentation.utils.errorToast
+import me.proton.core.presentation.utils.getUserMessage
 import me.proton.core.presentation.utils.hideKeyboard
 import me.proton.core.presentation.utils.onClick
 import me.proton.core.presentation.utils.onFailure
@@ -88,7 +89,8 @@ class SecondFactorActivity : AuthActivity<Activity2faBinding>(Activity2faBinding
                 is SecondFactorViewModel.State.Idle -> showLoading(false)
                 is SecondFactorViewModel.State.Processing -> showLoading(true)
                 is SecondFactorViewModel.State.AccountSetupResult -> onAccountSetupResult(it.result)
-                is SecondFactorViewModel.State.Error.Message -> onError(false, it.message)
+                is SecondFactorViewModel.State.Error.Message ->
+                    onError(false, it.error.getUserMessage(resources))
                 is SecondFactorViewModel.State.Error.Unrecoverable -> onUnrecoverableError(it.message)
             }.exhaustive
         }.launchIn(lifecycleScope)